requirements.txt
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.26.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6